LayoutTests/http/wpt/mediastream/transfer-mediastreamtrack-to-worker.html

<!DOCTYPE html>
<html>
<head>
<title>Transfer MediaStreamTrack to dedicated worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>

async function createWorker(script)
{
  script += "self.postMessage('ready');";
  const blob = new Blob([script], { type: 'text/javascript' });
  const url = URL.createObjectURL(blob);
  const worker = new Worker(URL.createObjectURL(blob));
  await new Promise(resolve => worker.onmessage = () => {
      resolve();
  });
  URL.revokeObjectURL(url);
  return worker;
}

promise_test(async test => {
   const stream = await navigator.mediaDevices.getUserMedia({ video: { width: 640, height: 480 } });

   const worker = await createWorker(`
       self.onmessage = async (event) => {
           const processor = new MediaStreamTrackProcessor({ track: event.data.track });
           const data = await processor.readable.getReader().read();
           self.postMessage(data);
           data.value.close();
       }
   `);
    const track = stream.getVideoTracks()[0].clone();
    worker.postMessage({ track }, [track]);
    
    const result = await new Promise(resolve => worker.onmessage = e => resolve(e.data));
    assert_false(result.done);

    test.add_cleanup(() => result.value.close());
    assert_equals(result.value.codedWidth, 640);
    assert_equals(result.value.codedHeight, 480);
}, "Transfer MediaStreamTrack to dedicated worker and use MediaStreamTrackProcessor");
</script>
</body>
</html>
